Consolidate the way a public type is calculated to DslObject
